Move method _get_type_name from H2OTypeError class into typechecks module, because it s much easier to update when the typechecks module gets updated
